Use v1 API in vertex tests and split vertex pipeline e2e tests to run them in parallel.
Also merged test_utils.BaseKubeflowV2Test and vertex_e2e_test.KubeflowV2E2ETestCase into a one base class in base_test_case.BaseKubeflowV2Test.

PiperOrigin-RevId: 395844000
